// types/nager.date.ts
/* eslint-disable unused-imports/no-unused-vars */
export interface NagerDatePublicHoliday {
  date: string;
  localName: string;
  name: string;
  countryCode: CountryCode;
  fixed: boolean;
  global: boolean;
  counties: null;
  launchYear: null;
  types: Type[];
}

export enum CountryCode {
  AD = 'AD',
  AL = 'AL',
  AM = 'AM',
  AR = 'AR',
  AT = 'AT',
  AU = 'AU',
  AX = 'AX',
  BA = 'BA',
  BB = 'BB',
  BE = 'BE',
  BG = 'BG',
  BJ = 'BJ',
  BO = 'BO',
  BR = 'BR',
  BS = 'BS',
  BW = 'BW',
  BY = 'BY',
  BZ = 'BZ',
  CA = 'CA',
  CH = 'CH',
  CL = 'CL',
  CN = 'CN',
  CO = 'CO',
  CR = 'CR',
  CU = 'CU',
  CY = 'CY',
  CZ = 'CZ',
  DE = 'DE',
  DK = 'DK',
  DO = 'DO',
  EC = 'EC',
  EE = 'EE',
  EG = 'EG',
  ES = 'ES',
  FI = 'FI',
  FO = 'FO',
  FR = 'FR',
  GA = 'GA',
  GB = 'GB',
  GD = 'GD',
  GE = 'GE',
  GG = 'GG',
  GI = 'GI',
  GL = 'GL',
  GM = 'GM',
  GR = 'GR',
  GT = 'GT',
  GY = 'GY',
  HK = 'HK',
  HN = 'HN',
  HR = 'HR',
  HT = 'HT',
  HU = 'HU',
  ID = 'ID',
  IE = 'IE',
  IM = 'IM',
  IS = 'IS',
  IT = 'IT',
  JE = 'JE',
  JM = 'JM',
  JP = 'JP',
  KR = 'KR',
  KZ = 'KZ',
  LI = 'LI',
  LS = 'LS',
  LT = 'LT',
  LU = 'LU',
  LV = 'LV',
  MA = 'MA',
  MC = 'MC',
  MD = 'MD',
  ME = 'ME',
  MG = 'MG',
  MK = 'MK',
  MN = 'MN',
  MS = 'MS',
  MT = 'MT',
  MX = 'MX',
  MZ = 'MZ',
  NA = 'NA',
  NE = 'NE',
  NG = 'NG',
  NI = 'NI',
  NL = 'NL',
  NO = 'NO',
  NZ = 'NZ',
  PA = 'PA',
  PE = 'PE',
  PG = 'PG',
  PL = 'PL',
  PR = 'PR',
  PT = 'PT',
  PY = 'PY',
  RO = 'RO',
  RS = 'RS',
  RU = 'RU',
  SE = 'SE',
  SG = 'SG',
  SI = 'SI',
  SJ = 'SJ',
  SK = 'SK',
  SM = 'SM',
  SR = 'SR',
  SV = 'SV',
  TN = 'TN',
  TR = 'TR',
  UA = 'UA',
  US = 'US',
  UY = 'UY',
  VA = 'VA',
  VE = 'VE',
  VN = 'VN',
  ZA = 'ZA',
  ZW = 'ZW',
}

export enum Type {
  Optional = 'Optional',
  Public = 'Public',
}